+++ services/value-store/projects/src/events/onObjectCreated.js
@@ -2,18 +2,19 @@ import AWS from 'aws-sdk'; // eslint-disable-line import/no-extraneous-dependenc
 
 import elasticsearch from 'elasticsearch';
 import connectionClass from 'http-aws-es';
-import through2 from 'through2';
-import split2 from 'split2';
 
 import MessengerFactory from '@eubfr/logger-messenger/src/lib/MessengerFactory';
 import { STATUS } from '@eubfr/logger-messenger/src/lib/status';
 
+// Helpers
 import deleteProjects from '../lib/deleteProjects';
-import SaveStream from '../lib/SaveStream';
+import saveToElasticSearch from '../lib/saveToElasticSearch';
+// import saveToKinesis from '../lib/saveToKinesis';
 
 export const handler = async (event, context, callback) => {
   const { API, INDEX, REGION, STAGE } = process.env;
 
+  // Checks ensuring proper functioning of the function
   if (!API || !INDEX || !REGION || !STAGE) {
     return callback(
       new Error(
@@ -21,153 +22,101 @@ export const handler = async (event, context, callback) => {
       )
     );
   }
-
-  /*
-   * Some checks here before going any further
-   */
   if (!event.Records) {
     return callback(new Error('No record'));
   }
-
   // Only work on the first record
   const snsRecord = event.Records[0];
-
   // Was the lambda triggered correctly? Is the file extension supported? etc.
   if (!snsRecord || snsRecord.EventSource !== 'aws:sns') {
     return callback(new Error('Bad record'));
   }
 
-  // Get Account ID from lambda function arn in the context
+  // Get keys
   const accountId = context.invokedFunctionArn.split(':')[4];
-  const sns = new AWS.SNS();
+  const s3record = JSON.parse(snsRecord.Sns.Message).Records[0];
+  const originalComputedKey = s3record.s3.object.key.replace('.ndjson', '');
 
-  // Insantiate clients
-  const client = elasticsearch.Client({
+  // Organize at a single place for helper functions
+  const usefulData = {
+    accountId,
+    originalComputedKey,
+    s3record,
+    env: process.env,
+  };
+
+  // Clients
+  const sns = new AWS.SNS();
+  const s3 = new AWS.S3();
+  const messenger = MessengerFactory.Create({ context });
+  const clientElasticSearch = elasticsearch.Client({
     host: `https://${API}`,
     apiVersion: '6.0',
     connectionClass,
     index: INDEX,
   });
 
-  const messenger = MessengerFactory.Create({ context });
-  const s3 = new AWS.S3();
+  // Organize at a single place for helper functions
+  const clients = { clientElasticSearch, messenger, s3, sns };
 
-  /*
-   * Extract information from the event
-   */
+  // Clear existing records
+  await deleteProjects({
+    client: clientElasticSearch,
+    index: INDEX,
+    key: s3record.s3.object.key,
+  });
 
-  // Extract S3 record
-  const s3record = JSON.parse(snsRecord.Sns.Message).Records[0];
-
-  // Get original computed key (without '.ndjson')
-  const originalComputedKey = s3record.s3.object.key.replace('.ndjson', '');
-
-  try {
+  // Define error handler
+  const handleError = async (e, cb) => {
     await messenger.send({
       message: {
         computed_key: originalComputedKey,
-        status_message: 'Preparing the upload to ElasticSearch...',
-        status_code: STATUS.PROGRESS,
+        status_message: e.message,
+        status_code: STATUS.ERROR,
       },
       to: ['logs'],
     });
 
-    const data = await s3
+    return cb(e);
+  };
+
+  try {
+    // Get information about the harmonized file
+    const fileData = await s3
       .headObject({
         Bucket: s3record.s3.bucket.name,
         Key: s3record.s3.object.key,
       })
       .promise();
 
-    await deleteProjects({
-      client,
-      index: INDEX,
-      key: s3record.s3.object.key,
-    });
+    // Add it to the box of useful data.
+    usefulData.fileData = fileData;
+
+    // If the file is less than 15MB, handle it directly
+    if (fileData.ContentLength <= 15000000) {
+      await messenger.send({
+        message: {
+          computed_key: originalComputedKey,
+          status_message: 'Start uploading to ElasticSearch',
+          status_code: STATUS.PROGRESS,
+        },
+        to: ['logs'],
+      });
+
+      return await saveToElasticSearch({ clients, usefulData, handleError });
+    }
 
     await messenger.send({
       message: {
         computed_key: originalComputedKey,
-        status_message: 'Start uploading to ElasticSearch',
+        status_message: 'Data is bigger than 15MB and will be streamed.',
         status_code: STATUS.PROGRESS,
       },
       to: ['logs'],
     });
 
-    // Prepare upload
-    const saveStream = new SaveStream({
-      objectMode: true,
-      client,
-      index: INDEX,
-    });
-
-    const handleError = async (e, cb) => {
-      await messenger.send({
-        message: {
-          computed_key: originalComputedKey,
-          status_message: e.message,
-          status_code: STATUS.ERROR,
-        },
-        to: ['logs'],
-      });
-
-      return cb(e);
-    };
-
-    const readStream = s3
-      .getObject({
-        Bucket: s3record.s3.bucket.name,
-        Key: s3record.s3.object.key,
-      })
-      .createReadStream();
-
-    return new Promise((resolve, reject) => {
-      readStream
-        .pipe(split2(JSON.parse))
-        .on('error', async e => handleError(e, reject))
-        .pipe(
-          through2.obj((chunk, enc, cb) => {
-            // Enhance item to save
-            const item = Object.assign(
-              {
-                computed_key: s3record.s3.object.key,
-                created_by: s3record.userIdentity.principalId, // which service created the harmonized file
-                last_modified: data.LastModified.toISOString(), // ISO-8601 date
-              },
-              chunk
-            );
-
-            return cb(null, item);
-          })
-        )
-        .on('error', async e => handleError(e, reject))
-        .pipe(saveStream)
-        .on('error', async e => handleError(e, reject))
-        .on('finish', async () => {
-          await messenger.send({
-            message: {
-              computed_key: originalComputedKey,
-              status_message: 'Results uploaded successfully, all went well.',
-              status_code: STATUS.INGESTED,
-            },
-            to: ['logs'],
-          });
-
-          const enrichmentTargetArn = `arn:aws:sns:${REGION}:${accountId}:${STAGE}-onEnrichmentRequested`;
-
-          await sns
-            .publish({
-              Message: JSON.stringify({
-                default: JSON.stringify(s3record.s3),
-              }),
-              MessageStructure: 'json',
-              TargetArn: enrichmentTargetArn,
-            })
-            .promise();
-
-          return resolve('Results uploaded successfully, all went well.');
-        });
-    });
+    // return await saveToKinesis({ clients, usefulData, handleError });
+    return callback(`Still to be implemented`);
   } catch (err) {
     await messenger.send({
       message: {
